# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyChangeOut": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "pending_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Until"
          }
        },
        "required": [
          "applied",
          "message"
        ],
        "title": "PolicyChangeOut",
        "type": "object"
      },
      "PolicyIn": {
        "properties": {
          "academic_level": {
            "anyOf": [
              {
                "enum": [
                  "sd",
                  "smp",
                  "sma",
                  "kuliah",
                  "profesional"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Level"
          },
          "allowed_upload_methods": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "photo",
                    "document"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Upload Methods"
          },
          "base_reward_seconds": {
            "anyOf": [
              {
                "maximum": 1800.0,
                "minimum": 15.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Reward Seconds"
          },
          "daily_caps": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "maxItems": 7,
                "minItems": 7,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Caps"
          },
          "daily_grants": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "maxItems": 7,
                "minItems": 7,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Grants"
          },
          "day_reset_hour": {
            "anyOf": [
              {
                "maximum": 23.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Reset Hour"
          },
          "essay_ratio": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Essay Ratio"
          },
          "idle_days_allowed": {
            "anyOf": [
              {
                "maximum": 14.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Days Allowed"
          },
          "locked_apps": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked Apps"
          },
          "question_language": {
            "anyOf": [
              {
                "enum": [
                  "id",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Language"
          },
          "questions_per_session": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Questions Per Session"
          }
        },
        "title": "PolicyIn",
        "type": "object"
      },
      "PolicyOut": {
        "properties": {
          "academic_level": {
            "title": "Academic Level",
            "type": "string"
          },
          "allowed_upload_methods": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Upload Methods",
            "type": "array"
          },
          "base_reward_seconds": {
            "title": "Base Reward Seconds",
            "type": "integer"
          },
          "daily_caps": {
            "items": {
              "type": "integer"
            },
            "title": "Daily Caps",
            "type": "array"
          },
          "daily_grants": {
            "items": {
              "type": "integer"
            },
            "title": "Daily Grants",
            "type": "array"
          },
          "day_reset_hour": {
            "title": "Day Reset Hour",
            "type": "integer"
          },
          "essay_ratio": {
            "title": "Essay Ratio",
            "type": "number"
          },
          "idle_days_allowed": {
            "title": "Idle Days Allowed",
            "type": "integer"
          },
          "locked_apps": {
            "items": {
              "type": "string"
            },
            "title": "Locked Apps",
            "type": "array"
          },
          "pending_weaken_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Weaken At"
          },
          "pending_weaken_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Weaken Payload"
          },
          "question_language": {
            "title": "Question Language",
            "type": "string"
          },
          "questions_per_session": {
            "title": "Questions Per Session",
            "type": "integer"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "subject_id",
          "questions_per_session",
          "base_reward_seconds",
          "essay_ratio",
          "academic_level",
          "question_language",
          "daily_caps",
          "daily_grants",
          "idle_days_allowed",
          "day_reset_hour",
          "allowed_upload_methods",
          "locked_apps"
        ],
        "title": "PolicyOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Kartu waktu bermain yang dicetak dari materi belajar",
    "title": "BrainXP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/policies/{subject_id}": {
      "get": {
        "operationId": "get_policy_policies__subject_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy",
        "tags": [
          "policy"
        ]
      },
      "put": {
        "operationId": "put_policy_policies__subject_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Policy",
        "tags": [
          "policy"
        ]
      }
    },
    "/policies/{subject_id}/locked-apps/{package}": {
      "delete": {
        "operationId": "unlock_app_policies__subject_id__locked_apps__package__delete",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package",
            "required": true,
            "schema": {
              "title": "Package",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unlock App",
        "tags": [
          "policy"
        ]
      },
      "post": {
        "operationId": "lock_app_policies__subject_id__locked_apps__package__post",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package",
            "required": true,
            "schema": {
              "title": "Package",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock App",
        "tags": [
          "policy"
        ]
      }
    },
    "/policies/{subject_id}/pending": {
      "delete": {
        "operationId": "cancel_pending_policies__subject_id__pending_delete",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Pending",
        "tags": [
          "policy"
        ]
      }
    },
    "/policies/{subject_id}/pending/apply": {
      "post": {
        "operationId": "apply_pending_policies__subject_id__pending_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Pending",
        "tags": [
          "policy"
        ]
      }
    },
...
  }
}